requirements.txt
Django==2.2.5
django-rest-framework-mongoengine==3.4.0
django-restframework==0.0.1
mongoengine==0.18.2
pymongo==3.9.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
